requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
python-dotenv>=1.0.0
scikit-learn>=1.4.0
sentence-transformers>=3.0.0
databricks-sql-connector>=3.0.0
torch>=2.0.0
